requirements.txt
streamlit
sockets
googlesearch-python
